Apollo server hooked up with fake schema
Had to make some significant changes
to get this working. Despite the docs,
it seems that it's required to use apollo
as an esm module, but the only way to get it
working seems to be to turn on esm
modules. Working for now, will keep playing
with it going forward and see what happens.